requirements.txt
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-iot-device
azure-digitaltwins-core
azure-eventhub
azure-functions